add gwcs to fits sip to importer, remove from orientation plugin
