+++ jdaviz/configs/imviz/plugins/orientation/orientation.vue
@@ -91,17 +91,6 @@
           />
         </v-row>
 
-        <v-row>
-          <plugin-switch
-            v-if="align_by_selected == 'WCS'"
-            :value.sync="gwcs_to_fits_sip"
-            label="Approximate GWCS with FITS SIP"
-            api_hint="plg.gwcs_to_fits_sip = "
-            :api_hints_enabled="api_hints_enabled"
-            hint="On future data loads, try to convert GWCS into FITS SIP for better performance (typical precision <0.1 pixels)."
-          />
-        </v-row>
-
         <v-row v-if="false">
           <plugin-switch
             :value.sync="wcs_use_fallback"
